+++ Assets/Scripts/Game_Scripts/Chroma_City_Neon_Racing/LevelSO.cs
@@ -13,5 +13,5 @@ public class LevelSO : ScriptableObject
     public int timePowerup;
     public int durationOfPowerups;
     public int timeLimit;
-    public int maxScore;
+    public float maxScore;
 }
